+++ src/index.ts
@@ -308,7 +308,7 @@ commands.set('who', async (interaction) => {
     // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
     // `\`IGNORE_BOTS\`: ${channel.config.IGNORE_BOTS}\n` +
     // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-    `\`IGNORE_EVERYONE_MENTIONS\`: ${channel.config.IGNORE_EVERYONE_MENTIONS}\n` +
+    // `\`IGNORE_EVERYONE_MENTIONS\`: ${channel.config.IGNORE_EVERYONE_MENTIONS}\n` +
     `\`LANGUAGE_MODEL\`: ${channel.config.LANGUAGE_MODEL}` +
     `\n\nSystem messages (${messages.length}): \n • ${messageList}`
 
@@ -384,7 +384,7 @@ commands.set('config', async (interaction) => {
       // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
       // `\`IGNORE_BOTS\`: ${channel.config.IGNORE_BOTS}\n` +
       // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-      `\`IGNORE_EVERYONE_MENTIONS\`: ${channel.config.IGNORE_EVERYONE_MENTIONS}\n` +
+      // `\`IGNORE_EVERYONE_MENTIONS\`: ${channel.config.IGNORE_EVERYONE_MENTIONS}\n` +
       `\`LANGUAGE_MODEL\`: ${channel.config.LANGUAGE_MODEL}`
   })
 })
@@ -597,11 +597,15 @@ commands.set('system', async (interaction) => {
 commands.set('help', async (interaction) => {
   await interaction.reply({
     content: 'Commands: \n' +
+      '`/who    `Gets information about the bot\'s configuration in this channel\n' +
       '`/tokens `Check how many tokens this server has left and links to buy more \n' +
       '`/config `Configure the bot for this channel (Admin only) \n' +
       '`/reset  `Reset the bot for this channel (Admin only)\n' +
       '`/system `Add a system message to the bot (Admin only) \n' +
-      '`/help   `Show this message',
+      '`/help   `Show this message' +
+      '\n\nRead more on the [website](https://discordai.chat)' +
+      '\n\nFor help, please join the ' +
+      '[support server](https://discord.gg/b8FKqtjRX5)',
     ephemeral: true
   })
 })