+++ src/logger/index.ts
@@ -8,16 +8,18 @@
 
 import { LogEntry } from './LogEntry'
 
-export const log = ({ message, guildId, channelId, level }:
+export const log = ({ message, guildId, channelId, level, userId }:
 {
   message: string
   guildId?: string
   channelId?: string
-  level?: 'info' | 'warn' | 'error' }
+  level?: 'info' | 'warn' | 'error'
+  userId?: string
+}
 ): void => {
   // print to console
   console.log(`${guildId ?? '<missing guild>'}:${channelId ?? '<missing channel>'}: ${message}`)
   // store in database
-  const logEntry = new LogEntry(message, guildId, channelId, level)
+  const logEntry = new LogEntry(message, guildId, channelId, level, userId)
   void logEntry.save()
 }