+++ src/messages/Guild.ts
@@ -4,7 +4,6 @@ import { type ChannelConfig } from './ChannelConfig'
 import fs from 'fs'
 import { DEFAULT_MODERATION_VIOATION_RESPONSE, DISCLAIMER } from '../env'
 import { mongoClient } from '../mongo/MongoClient'
-import { log } from '../logger'
 
 export const DEFAULT_GUILD_CONFIG: ChannelConfig = {
   ERROR_RESPONSE: 'Sorry, there was an error. Please try again later.',
@@ -49,11 +48,6 @@ export class Guild {
 
   async getChannel (id: string): Promise<Channel | undefined> {
     if (!this.channels.has(id)) {
-      log({
-        guildId: this.id,
-        channelId: id,
-        message: 'Channel not found'
-      })
       return undefined
     }
     return this.channels.get(id)